Fixes device commands with 0 valued parameters.
When sending device commands, explicitly compare parameter values to null. This
allows command parameters that would otherwise implicitly evaluate to false in
conditionals (e.g. 0 value integers.)

Prior to this change, for example, a 'setLevel(0)' command would be sent to the
device handler as 'setLevel()' and not all devices (e.g. window coverings) are
able to gracefully handle the elided parameter value.